добавляю удаление старых данных

+++ version_1/learning_try_1/learning_app/views.py
@@ -77,6 +77,7 @@ def input_one(request):
 
 
 def delete_old(request):
-    two_days_ago = datetime.date.today() - datetime.timedelta(10)
-    Vacancy.objects.filter(data_added__lte=two_days_ago).delete()
-    return render(request, 'learning_app/home.html', 'delete')
+    if request.user.is_authenticated:
+        two_days_ago = datetime.date.today() - datetime.timedelta(10)
+        Vacancy.objects.filter(owner=request.user).filter(data_added__lte=two_days_ago).delete()
+        return render(request, 'learning_app/home.html', 'delete')
